- 局部重构k8s start & new k8s 框架
- ioutil -> os | io
- defer unlock
- 链接文件相对于.ide.yaml 文件所在目录 && 解析链接k8s yaml 错误
- 不同 type 和 subtype 模板应该是新的工作区 && new 工作区名称为空，默认使用type && 开发浏览器后命令行exit问题
- new local error: repository does not exist && cmd 获取工作区信息方法移动到common
- 中获取 开发容器名称失败 && 本地相同目录再次运行的时候，提示 actual git repo url is null! && 工作区的名称没有加随机字符串
- error: current-context is not set
- server 运行是报错：actual git repo url is null!;
- chmod: cannot access '/home/smartide/.ssh': No such file or directory

Related work items: #1756

+++ cli/cmd/remove/k8s.go
@@ -1,7 +1,7 @@
 /*
  * @Date: 2022-06-07 14:02:38
  * @LastEditors: Jason Chen
- * @LastEditTime: 2022-09-09 16:22:21
+ * @LastEditTime: 2022-10-28 15:05:27
  * @FilePath: /cli/cmd/remove/k8s.go
  */
 /*
@@ -49,8 +49,10 @@ func RemoveK8s(k8sUtil k8s.KubernetesUtil, workspaceInfo workspace.WorkspaceInfo
 		return err
 	}
 	repoName := common.GetRepoName(workspaceInfo.GitCloneRepoUrl)
-	filePath := common.PathJoin(home, ".ide", repoName)
-	os.RemoveAll(filePath)
+	if repoName != "" {
+		filePath := common.PathJoin(home, ".ide", repoName)
+		os.RemoveAll(filePath)
+	}
 
 	//remove config note from .ssh/config file
 	workspaceInfo.RemoveSSHConfig()
